Fixed the normal list item delete swipe, abstracted gesture to its own component, and started on inset list item swipe to delete
