+++ habit-tracker-frontend/src/global.scss
@@ -121,7 +121,7 @@ ion-checkbox {
 
 .swipe-icon-row {
     background-color: var(--ion-color-danger);
-    z-index: -1;
+    z-index: 1;
     grid-column: 1;
     grid-row: 1;
     display: grid;
@@ -136,4 +136,17 @@ ion-checkbox {
     flex-direction: row;
     align-items: center;
     justify-content: space-between;
+}
+
+.dtl-list-checkbox-row {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.dtl-swipe-icon {
+    height: 20px;
+    width: 20px;
+    padding-left: 10px;
+    padding-right: 10px;
 }
